fix: Modal popup opens at wrong time
